+++ packages/sampler/stories/qa-stories/ExpandableList.js
@@ -40,7 +40,7 @@ storiesOf('ExpandableList')
 		'with children length update',
 		() => (
 			<ExpandableListChildrenLengthUpdate
-				autoCloseOnSelect={boolean('autoCloseOnSelect', false)}
+				closeOnSelect={boolean('closeOnSelect', false)}
 				disabled={boolean('disabled', false)}
 				noAutoClose={boolean('noAutoClose', false)}
 				noLockBottom={boolean('noLockBottom', false)}